Creating API verification, fixing PATCH method

+++ controllers/recipes.js
@@ -1,7 +1,7 @@
-// Importing the Recipe model from ../models/Recipe
+// Importing the Recipe model from ../models/Recipe.
 const Recipe = require('../models/Recipe');
 
-// getAll function retrieves all recipes from the database
+// getAll function retrieves all recipes from the database.
 async function getAll(req, res) {
   try {
     const recipes = await Recipe.find({});
@@ -13,7 +13,7 @@ async function getAll(req, res) {
   }
 }
 
-// getByCategory function retrieves all recipes belonging to the category from the database
+// getByCategory function retrieves all recipes belonging to the category from the database.
 async function getByCategory(req, res) {
   const category = req.params.category;
   try {
@@ -26,7 +26,7 @@ async function getByCategory(req, res) {
   }
 }
 
-// getById function retrieves recipe by id from the database
+// getById function retrieves recipe by id from the database.
 async function getById(req, res) {
   const id = req.params.id;
   try {
@@ -39,10 +39,24 @@ async function getById(req, res) {
   }
 }
 
-// Creating a new recipe based on data received in req.body
+// Creating a new recipe based on data received in req.body.
 async function create(req, res) {
   try {
-    // req.body will have the info that the user filled out on the frontend
+    const key = req.query['api-key'];
+
+    // Check for the absence of a key.
+    if (!key) {
+      res.status(400).send('API Key Required');
+      return;
+    }
+
+    // Check for key validity.
+    if (process.env.API_KEY !== key) {
+      res.status(401).send('Invalid API Key');
+      return;
+    }
+
+    // req.body will have the info that the user filled out on the frontend.
     const createdRecipe = await Recipe.create(req.body);
     if (createdRecipe) {
       res.status(201).json(createdRecipe);
@@ -55,7 +69,21 @@ async function create(req, res) {
 // Updating a recipe in the database based on the id passed in req.params.id.
 async function update(req, res) {
   try {
-    const updatedRecipe = await Recipe.findByIdAndUpdate(req.params.id);
+    const key = req.query['api-key'];
+
+    // Check for the absence of a key.
+    if (!key) {
+      res.status(400).send('API Key Required');
+      return;
+    }
+
+    // Check for key validity.
+    if (process.env.API_KEY !== key) {
+      res.status(401).send('Invalid API Key');
+      return;
+    }
+
+    const updatedRecipe = await Recipe.findByIdAndUpdate(req.params.id, req.body);
     if (updatedRecipe) {
       res.status(200).json(updatedRecipe);
     }
@@ -67,6 +95,20 @@ async function update(req, res) {
 // Deleting a recipe from the database based on the id passed in req.params.id.
 async function remove(req, res) {
   try {
+    const key = req.query['api-key'];
+
+    // Check for the absence of a key.
+    if (!key) {
+      res.status(400).send('API Key Required');
+      return;
+    }
+
+    // Check for key validity.
+    if (process.env.API_KEY !== key) {
+      res.status(401).send('Invalid API Key');
+      return;
+    }
+
     const deletedRecipe = await Recipe.findByIdAndDelete(req.params.id);
     if (deletedRecipe) {
       res.status(204).json(deletedRecipe);
@@ -76,9 +118,7 @@ async function remove(req, res) {
   }
 }
 
-//Exporting the getAll, create, and destroy functions 
-//so they can be used by other parts of the application (routes)
-
+// Exporting functions to be used by other parts of the application (routes).
 module.exports = {
   getAll,
   getByCategory,
